changed validator due to previous logic being redudand
